shrank down unnecessary utility class MethodDispatchResolver. Removed unnecessary tests. Moved utility methods to the actual class using them. moved tests.
