+++ sootup.callgraph/src/main/java/sootup/callgraph/RapidTypeAnalysisAlgorithm.java
@@ -174,15 +174,14 @@ public class RapidTypeAnalysisAlgorithm extends AbstractCallGraphAlgorithm {
       Set<MethodSignature> concreteCallTargets =
           implAndOverrides.stream()
               .map(
-                  methodSignature ->
-                      MethodDispatchResolver.resolveConcreteDispatch(view, methodSignature))
+                  methodSignature ->resolveConcreteDispatch(view, methodSignature))
               .filter(Optional::isPresent)
               .map(Optional::get)
               .collect(Collectors.toSet());
 
       // add the concrete of the targetMethod if the class is instantiated
       if (targetMethodClassIsInstantiated) {
-        MethodDispatchResolver.resolveConcreteDispatch(view, targetMethodSignature)
+        resolveConcreteDispatch(view, targetMethodSignature)
             .ifPresent(concreteCallTargets::add);
       }
 
@@ -239,8 +238,7 @@ public class RapidTypeAnalysisAlgorithm extends AbstractCallGraphAlgorithm {
           if (newEdges != null) {
             newEdges.forEach(
                 call -> {
-                  MethodSignature concreteTarget =
-                      MethodDispatchResolver.resolveConcreteDispatch(view, call.target)
+                  MethodSignature concreteTarget =resolveConcreteDispatch(view, call.target)
                           .orElse(null);
                   if (concreteTarget == null) {
                     return;
