+++ sootup.callgraph/src/test/resources/callgraph/ConcreteDispatch/B.java
