--- sootup.tests/src/test/resources/methoddispatchresolver/AbstractDispatch/A.java
@@ -1,8 +1,0 @@
-
-/** @author: Hasitha Rajapakse **/
-
-public class A {
-    public void method(){
-        int num = 5;
-    }
-}
