--- sootup.tests/src/test/resources/methoddispatchresolver/AbstractDispatch/AbstractClass.java
@@ -1,6 +1,0 @@
-
-/** @author: Hasitha Rajapakse **/
-
-public abstract class AbstractClass{
-    public abstract void method();
-}
