--- sootup.tests/src/test/resources/methoddispatchresolver/AbstractDispatch/B.java
@@ -1,5 +1,0 @@
-
-/** @author: Hasitha Rajapakse **/
-
-public class B extends AbstractClass {
-}
