--- sootup.tests/src/test/resources/methoddispatchresolver/AbstractDispatch/C.java
@@ -1,8 +1,0 @@
-
-/** @author: Hasitha Rajapakse **/
-
-public class C extends AbstractClass {
-    public void method(){
-        int num = 20;
-    }
-}
