--- sootup.tests/src/test/resources/methoddispatchresolver/AllDispatch/A.java
@@ -1,7 +1,0 @@
-
-/** @author: Jonas Klauke **/
-
-public class A {
-    public void method(){
-    }
-}
