--- sootup.tests/src/test/resources/methoddispatchresolver/AllDispatch/AbstractClass.java
@@ -1,6 +1,0 @@
-
-/** @author: Jonas Klauke **/
-
-public abstract class AbstractClass extends A{
-    public abstract void method();
-}
