--- sootup.tests/src/test/resources/methoddispatchresolver/AllDispatch/B.java
@@ -1,5 +1,0 @@
-
-/** @author: Jonas Klauke **/
-
-public class B extends AbstractClass {
-}
