--- sootup.tests/src/test/resources/methoddispatchresolver/AllDispatch/C.java
@@ -1,8 +1,0 @@
-
-/** @author: Jonas Klauke **/
-
-public class C extends A {
-    public void method(){
-        int num = 20;
-    }
-}
