--- sootup.tests/src/test/resources/methoddispatchresolver/CovarianceDispatch/Subclass.java
@@ -1,8 +1,0 @@
-
-/** @author: Jonas Klauke **/
-
-public class Subclass extends Superclass {
-    public Superclass method(){
-        return new Superclass();
-    }
-}
