--- sootup.tests/src/test/resources/methoddispatchresolver/CovarianceDispatch/Superclass.java
@@ -1,8 +1,0 @@
-
-/** @author: Jonas Klauke **/
-
-public class Superclass{
-    public Object method(){
-        return new Object();
-    }
-}
